backend/server/requirements.txt
alembic==1.4.2
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
firebase-admin==4.4.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==1.0.0
mysqlclient==2.0.3
passlib==1.7.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-dotenv==0.10.1
pytz==2020.4
requests==2.24.0
rsa==4.6
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.17
uritemplate==3.0.1
urllib3==1.25.11
uuid==1.30
Werkzeug==1.0.1
WTForms==2.3.1